feat: accum pow
